Factor code into src folder
