--- projects/playwright-runner/index.spec.ts
@@ -1,5 +1,0 @@
-import test from "node:test";
-
-test("handler", () => {
-    
-});
