--- projects/playwright-runner/model.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const AuditPageOptionsSchema = z.object({
-  url: z.string(),
-});
-
-export type AuditPageOptions = z.infer<typeof AuditPageOptionsSchema>;
