+++ projects/playwright-runner/src/index.ts
@@ -1,0 +1,4 @@
+import { auditPage } from "./audit.ts";
+import { createHandler } from "./handler.ts";
+
+export const handler = createHandler(auditPage);
