+++ projects/playwright-runner/src/model/request.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const AuditPageRequestSchema = z.object({
+  url: z.string(),
+});
+
+export type AuditPageRequest = z.infer<typeof AuditPageRequestSchema>;
